got rid of excessive convertation to double while round up majority number

+++ service-node/src/main/scala/org/enterprisedlt/fabric/service/node/Util.scala
@@ -90,7 +90,7 @@ object Util {
         parties.filter(e => e.role == role).map(_.mspId)
     }
 
-    def roundUpMajorityNumber(participantNumber: Int): Int = math.ceil(participantNumber.toDouble).toInt
+    def roundUpMajorityNumber(participantNumber: Int): Int = math.ceil(participantNumber).toInt
 
     //=========================================================================
     private def makeMSPPrincipal(memberName: String): MSPPrincipal = {
